Allow rules to be configured in the plugin block
It's not possible to come up with a list of which attributes on which
resources should include the workspace, because this depends on the
project.

In AWS for example, if each workspace creates its own VPC then the VPC
name needs to include the workspace, but things like security groups
within the VPC don't need to. On the other hand, if each workspace
exists within the same VPC then the security groups do need unique
names.

Letting the user configure the resources also decouples this plugin from
any particular provider - although I intend to use it mostly for AWS, it
should work with Azure, GCP, and every other provider