Simplify handling of non-default workspaces
Originally I thought that I could sneakily override TF_WORKSPACE while
this plugin ran, and have tflint respect that. Of course that doesn't
work, because the plugin is a child process of tflint, so setting an
environment in the one doesn't affect the other.

Instead, it's simpler just to ask the user to set TF_WORKSPACE before
they run tflint, if they don't want to test the default workspace.

This leaves a slightly suboptimal situation where the linter can pass if
TF_WORKSPACE is set, but fail otherwise if the config being tested has a
special case for "default" (which is fairly common). For now, 🤷‍♀️